requirements.txt
Flask==1.0.2
flask-rebar==1.4.0
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.4.0
python-dotenv==0.10.1
psycopg2==2.7.7
